moved card to the entity folder
Took 9 minutes
